Select color from ColorPicker using image instead of pod
